Improvements and refactoring
