+++ src/components/Pagination/style.module.css
@@ -28,7 +28,7 @@
 }
 
 .linkActive {
-  background-color: var(--dark-bg-color);
+  background-color: var(--accent-color);
   color: #fff;
 }
