+++ src/components/product-updates/ProductUpdate/Component.astro
@@ -1,0 +1,50 @@
+---
+import type { FragmentOf } from 'gql.tada';
+import { readFragment } from 'gql.tada';
+import { Image } from '~/components/blocks/Image';
+import { InternalVideo } from '~/components/blocks/InternalVideo';
+import { Prose } from '~/components/Prose';
+import { Text } from '~/components/structuredText/Text';
+import { withAllComponents } from '~/lib/datocms/structuredText';
+import { formatDate } from '~/lib/formatDate';
+import { ProductUpdateFragment } from './graphql';
+import s from './style.module.css';
+
+interface Props {
+  productUpdate: FragmentOf<typeof ProductUpdateFragment>;
+}
+
+const { productUpdate: maskedProductUpdate } = Astro.props;
+
+const productUpdate = readFragment(ProductUpdateFragment, maskedProductUpdate);
+---
+
+<article class={s.post}>
+  <div class={s.categories}>
+    {
+      productUpdate.categories.map((cat) => (
+        <span class={s.category} style={{ backgroundColor: cat.color?.hex }}>
+          {cat.name}
+        </span>
+      ))
+    }
+  </div>
+
+  <h2 class={s.title}>
+    <a href={`/product-updates/${productUpdate.slug}`}>{productUpdate.title}</a>
+  </h2>
+
+  <div class={s.info}>{formatDate(productUpdate._firstPublishedAt!)}</div>
+
+  <Prose>
+    <div class={s.body}>
+      <Text
+        data={productUpdate.content}
+        blockComponents={withAllComponents(productUpdate.content!.blocks, {
+          InternalVideoRecord: InternalVideo,
+          ImageRecord: Image,
+        })}
+      />
+    </div>
+  </Prose>
+</article>
