+++ src/components/product-updates/ProductUpdate/index.ts
@@ -1,0 +1,1 @@
+export { default as ProductUpdate } from './Component.astro';
