+++ src/components/product-updates/ProductUpdate/style.module.css
@@ -1,19 +1,21 @@
 .post {
-  border-top: 2px solid var(--border-color);
-  font-family: var(--font-serif);
-  letter-spacing: -0.02em;
-  margin: 0 auto;
-  max-width: 750px;
-  padding: rfs(90px) 0;
+  max-width: 900px;
 
-  &:first-child {
-    margin-top: 0;
+  @media screen and (min-width: 750px) {
+    padding: rfs(90px);
+    border-radius: 10px;
+    margin: rfs(50px) auto;
+    box-shadow: var(--box-shadow);
   }
 
-  &:last-child {
-    border-bottom: 2px solid var(--border-color);
-    margin-bottom: rfs(90px);
-    padding-bottom: rfs(90px);
+  @media screen and (max-width: 749px) {
+    border-top: 2px solid var(--border-color);
+    padding: rfs(90px) 0;
+
+    &:first-child {
+      padding-top: 0;
+      border-top: 0;
+    }
   }
 }
 
@@ -58,6 +60,5 @@
 }
 
 .body {
-  font-size: rfs(19px);
-  line-height: rfs(1.6);
+  line-height: rfs(1.5);
 }
