--- src/layouts/product-updates/index.ts
@@ -1,1 +1,0 @@
-export { default as ProductUpdatesLayout } from './Component.astro';
