--- src/pages/product-updates/[slug]/_style.module.css
@@ -1,63 +1,0 @@
-.post {
-  border-top: 2px solid var(--border-color);
-  font-family: var(--font-serif);
-  letter-spacing: -0.02em;
-  margin: 0 auto;
-  max-width: 750px;
-  padding: rfs(90px) 0;
-
-  &:first-child {
-    margin-top: 0;
-  }
-
-  &:last-child {
-    border-bottom: 2px solid var(--border-color);
-    margin-bottom: rfs(90px);
-    padding-bottom: rfs(90px);
-  }
-}
-
-.categories {
-  margin-bottom: 30px;
-}
-
-.category {
-  border-radius: 4px;
-  box-shadow: var(--box-shadow);
-  color: white;
-  display: inline-block;
-  font-family: var(--font-sans);
-  font-size: 0.8em;
-  font-weight: bold;
-  margin-right: 10px;
-  padding: 6px 12px 4px;
-}
-
-.title {
-  font-family: var(--font-sans);
-  font-size: rfs(35px);
-  letter-spacing: -0.05em;
-  line-height: 1.2;
-  margin: 0;
-
-  a {
-    color: inherit;
-    text-decoration: none;
-
-    &:hover {
-      text-decoration: underline;
-    }
-  }
-}
-
-.info {
-  color: var(--light-body-color);
-  font-family: var(--font-sans);
-  margin-bottom: rfs(40px);
-  margin-top: 10px;
-}
-
-.body {
-  font-size: rfs(19px);
-  line-height: rfs(1.6);
-}
