+++ src/pages/product-updates/index.astro
@@ -1,5 +1,5 @@
 ---
-import ProductUpdatesLayout from '~/layouts/product-updates/Component.astro';
+import ProductUpdates from './p/[pageIndex]/index.astro';
 ---
 
-<ProductUpdatesLayout />
+<ProductUpdates pageIndex={1} />
